Fix declaration of HAL_Time_SysTicksToTime for STM32

+++ targets/AzureRTOS/_common/targetHAL_Time.cpp
@@ -6,7 +6,7 @@
 #include <nanoHAL.h>
 
 // Converts CMSIS sysTicks to .NET ticks (100 nanoseconds)
-uint64_t HAL_Time_SysTicksToTime(unsigned int sysTicks) 
+uint64_t HAL_Time_SysTicksToTime(uint64_t sysTicks) 
 {
     // convert to microseconds from ST HAL SysTicks
     // need to convert from microseconds to 100 nanoseconds
@@ -16,7 +16,7 @@ uint64_t HAL_Time_SysTicksToTime(unsigned int sysTicks)
 // because HAL_Time_SysTicksToTime needs to be called from C we need a proxy to allow it to be called from 'C' code
 extern "C" {
     
-    uint64_t HAL_Time_SysTicksToTime_C(unsigned int sysTicks)
+    uint64_t HAL_Time_SysTicksToTime_C(uint64_t sysTicks)
     {
         return HAL_Time_SysTicksToTime(sysTicks);
     }
